Version bump.
The quad that is used for rendering is now defined in Pass. This means
that all passes that inherit from Pass don't have to define it anymore.
The scene, the camera and the quad field may all be set to null in case
you don't want them to be created.